Added connect retry loop for database connection
The MySQL local development container takes a while to fully start.
When the backend and database together with docker compose, the
backend would come online first, and fail it's connection to the
database.

Lets add a configurable connect retry loop to the backend to
check multiple times if the database is available, before shutting
down the server.